# bufio: optimize bufio.Reader.ReadString to avoid an allocation and copy